cursor-finish-version/requirements.txt
numpy>=1.24.3
torch>=2.0.1
gymnasium>=0.28.1
wandb>=0.15.4
tqdm>=4.65.0
pyyaml>=6.0
bayesian-optimization>=1.4.0 
wandb
matplotlib